+++ src/Plutarch/Orphans.hs
@@ -34,6 +34,7 @@ import PlutusLedgerApi.V2 (
   Credential (PubKeyCredential, ScriptCredential),
   CurrencySymbol (CurrencySymbol),
   Data (I, List),
+  Datum,
   LedgerBytes (LedgerBytes),
   MintingPolicy (MintingPolicy),
   POSIXTime (POSIXTime),
@@ -128,6 +129,12 @@ instance
       . bytes
       . coerce @(AsBase16Bytes a) @LedgerBytes
 
+  toEncoding =
+    Aeson.toEncoding @Text
+      . encodeByteString
+      . bytes
+      . coerce @(AsBase16Bytes a) @LedgerBytes
+
 -- @ since 3.6.1
 instance
   (Coercible LedgerBytes a) =>
@@ -153,6 +160,13 @@ instance (Serialise a) => Aeson.ToJSON (AsBase16Codec a) where
       . serialise @a
       $ x
 
+  toEncoding (AsBase16Codec x) =
+    Aeson.toEncoding @Text
+      . encodeByteString
+      . toStrict
+      . serialise @a
+      $ x
+
 -- @ since 3.6.1
 instance (Serialise a) => Aeson.FromJSON (AsBase16Codec a) where
   parseJSON v =
@@ -163,6 +177,12 @@ instance (Serialise a) => Aeson.FromJSON (AsBase16Codec a) where
         . deserialiseOrFail
         . fromStrict
         . encodeUtf8
+
+-- @since X.Y.Z
+deriving via (AsBase16Codec Datum) instance Aeson.ToJSON Datum
+
+-- @since X.Y.Z
+deriving via (AsBase16Codec Datum) instance Aeson.FromJSON Datum
 
 -- @ since 3.6.1
 deriving via (AsBase16Bytes TxId) instance Aeson.ToJSON TxId