+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 3.18.1 -- 2022-12-02
+
+### Added
+
+- `AsBase16Codec`-based `ToJSON`, `FromJSON` instances for `Datum` 
+
+### Modified
+
+- Added `toEncoding` implementation for `AsBase16Bytes` & `AsBase16Codec` `ToJSON` instances
+
 ## 3.18.0 -- 2022-11-30
 
 ### Modified